+++ drivers/CoreVideo/include/CoreGraphics.hpp
@@ -7,7 +7,7 @@
 
 #include "CGColor.hpp"
 #include "LKCoreLL.h"
-#include "interface/CoreDMA2D.hpp"
+#include "interface/DMA2D.hpp"
 #include "interface/CoreGraphics.hpp"
 #include "internal/corevideo_config.h"
